+++ pytest_inmanta/core.py
@@ -16,10 +16,8 @@ limitations under the License.
 Contact: code@inmanta.com
 """
 
+import importlib.metadata
 from typing import Optional
-
-import pkg_resources
-from pkg_resources import DistributionNotFound
 
 from packaging import version
 
@@ -29,10 +27,8 @@ Version of the inmanta-core package. None if it is not installed.
 """
 
 try:
-    CORE_VERSION = version.Version(
-        pkg_resources.get_distribution("inmanta-core").version
-    )
-except DistributionNotFound:
+    CORE_VERSION = version.Version(importlib.metadata.version("inmanta-core"))
+except importlib.metadata.PackageNotFoundError:
     CORE_VERSION = None