+++ tests/conftest.py
@@ -24,7 +24,6 @@ import tempfile
 from pathlib import Path
 from typing import Iterator, Optional
 
-import pkg_resources
 import pytest
 
 import pytest_inmanta.plugin
@@ -58,7 +57,6 @@ def deactive_venv():
     old_path_hooks = sys.path_hooks.copy()
     old_pythonpath = os.environ.get("PYTHONPATH", None)
     old_os_venv: Optional[str] = os.environ.get("VIRTUAL_ENV", None)
-    old_working_set = pkg_resources.working_set
 
     yield
 
@@ -72,7 +70,6 @@ def deactive_venv():
     sys.path_hooks.extend(old_path_hooks)
     # Clear cache for sys.path_hooks
     sys.path_importer_cache.clear()
-    pkg_resources.working_set = old_working_set
     # Restore PYTHONPATH
     if old_pythonpath is not None:
         os.environ["PYTHONPATH"] = old_pythonpath